Allow earning more than 255 coins from money weather

+++ src/team_feed/mod.rs
@@ -64,7 +64,7 @@ pub enum ParsedTeamFeedEventText<S> {
     },
     Prosperous {
         team: EmojiTeam<S>,
-        income: u8
+        income: u32
     },
     DonatedToLottery {
         team_name: S,
